#23 - Implement "play:clean" mojo - temporary fix to exclude "db/evolutions" from deleting when "db" directory is deleted - cont.
git-svn-id: http://maven-play-plugin.googlecode.com/svn/trunk/plugin@8557 9c2ad861-7886-cb7c-dcd7-2b8d41cd5de0

+++ play-maven-plugin/src/main/java/com/google/code/play/PlayCleanMojo.java
@@ -162,10 +162,13 @@ public class PlayCleanMojo
         {
             //TODO - this is temporary solution to avoid deleting "db/evolutions" subdirectory, improve it
             File dbDir = new File( baseDir, "db" );
-            deleteDirectory( new File( dbDir, "h2" ) );
-            if ( dbDir.list().length == 0 ) // empty directory
+            if ( dbDir.exists() )
             {
-                deleteDirectory( dbDir );
+                deleteDirectory( new File( dbDir, "h2" ) );
+                if ( dbDir.list().length == 0 ) // empty directory
+                {
+                    deleteDirectory( dbDir );
+                }
             }
         }
